RPTrees

Create ensembles of random projection trees in Julia.

Functions

RPTrees.RPForest — Method

RPForest(data::Array{T, 2}, maxdepth::Int, ntrees::Int) where T -> ensemble

Constructor for ensemble of sparse random projection trees with voting. Follows the implementation outlined in:

Fast Nearest Neighbor Search through Sparse Random Projections and Voting. Ville Hyvönen, Teemu Pitkänen, Sotirios Tasoulis, Elias Jääsaari, Risto Tuomainen, Liang Wang, Jukka Ilmari Corander, Teemu Roos. Proceedings of the 2016 IEEE Conference on Big Data (2016)

RPTrees.traverse_to_leaves — Method

traverse_to_leaves(rpf::RPForest{T}, x::Array{T, 2}) where T -> leaf_idxs

Route data point x down to a leaf node each tree and return and array of indexes of the data stored in each corresponding leaf node