RPTrees

Create ensembles of random projection trees in Julia.

Functions

RPTrees.RPForest — Method

RPForest(data::Array{T, 2}, maxdepth::Int, ntrees::Int) where T -> ensemble

Constructor for ensemble of sparse random projection trees with voting. Follows the implementation outlined in:

Fast Nearest Neighbor Search through Sparse Random Projections and Voting. Ville Hyvönen, Teemu Pitkänen, Sotirios Tasoulis, Elias Jääsaari, Risto Tuomainen, Liang Wang, Jukka Ilmari Corander, Teemu Roos. Proceedings of the 2016 IEEE Conference on Big Data (2016)

RPTrees.knngraph — Method

knngraph(rpf::RPForest{T}, k::Int, vote_cutoff; vote_cutoff::Int=1, ne_iters::Int=0, gtype::G) where {T, G} -> g

Returns a graph with rpf.npoints node and k * rpf.npoints edges datapoints conneceted to nearest neighbors

Parameters

rpf: random forest of the desired data

k: the desired number of nearest neighbors

vote_cutoff: signifies how many "votes" a point needs in order to be included

in a linear search through leaf nodes. Increasing vote_cutoff speeds up the algorithm but may reduce accuracy. Defaults to 1

ne_iters: assigns the number of iterations of neighbor exploration to use. Defaults to zero.

Neighbor exploration is a way to increse knn-graph accuracy.

gtype is the type of graph to construct. Defaults to SimpleDiGraph

RPTrees.approx_knn — Method

approx_knn(rpf::RPForest{T}, q::Array{T, 2}, k::Int; vote_cutoff=1) where T -> knn_idx

For a query point q, find the approximate k nearest neighbors from the data stored in the the RPForest. The vote_cutoff parameter signifies how many "votes" a point needs in order to be included in a linear search. Increasing vote_cutoff speeds up the algorithm but may reduce accuracy.

RPTrees.allknn — Method

allknn(rpf::RPForest{T}, k::Int; vote_cutoff::Int=1, ne_iters::Int=0) where T -> approxnn

Returns a rpf.npoints by k array of approximate nearest neighbor indexes. That is, approxnn[i,:] contains the indexes of the k nearest neighbors of rpf.data[:, i].

The ne_iters assigns the number of iterations of neighbor exploration to use. Neighbor exploration is an inexpensive way to increase accuracy.

The vote_cutoff parameter signifies how many "votes" a point needs in order to be included in a linear search. Increasing vote_cutoff speeds up the algorithm but may reduce accuracy.

RPTrees.traverse_to_leaves — Method

traverse_to_leaves(rpf::RPForest{T}, x::Array{T, 2}) where T -> leaf_idxs

Route data point x down to a leaf node each tree and return and array of indexes of the data stored in each corresponding leaf node